# back-end/Docs/api-definition.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "WebApi.Presentation",
    "version": "1.0"
  },
  "paths": {
    "/citizenship": {
      "get": {
        "tags": ["Citizenship"],
        "summary": "Gets all citizenship information.",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CitizenshipResponseDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CitizenshipResponseDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CitizenshipResponseDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/city": {
      "get": {
        "tags": ["City"],
        "summary": "Retrieves a list of all cities.",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CityResponseDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CityResponseDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CityResponseDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/disability": {
      "get": {
        "tags": ["Disability"],
        "summary": "Retrieves a list of all disabilities.",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DisabilityResponseDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DisabilityResponseDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DisabilityResponseDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/marital-status": {
      "get": {
        "tags": ["MaritalStatus"],
        "summary": "Retrieves a list of all marital statuses.",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MaritalStatusResponseDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MaritalStatusResponseDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MaritalStatusResponseDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/personal-info": {
      "post": {
        "tags": ["PersonalInfo"],
        "summary": "Creates personal information.",
        "requestBody": {
          "description": "The personal information to create.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonalInfoRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonalInfoRequestDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PersonalInfoRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PersonalInfoResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonalInfoResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonalInfoResponseDto"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["PersonalInfo"],
        "summary": "Gets all personal information.",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PersonalInfoResponseDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PersonalInfoResponseDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PersonalInfoResponseDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/personal-info/{id}": {
      "get": {
        "tags": ["PersonalInfo"],
        "summary": "Gets personal information by ID.",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "The ID of the personal information to retrieve.",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PersonalInfoResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonalInfoResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonalInfoResponseDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["PersonalInfo"],
        "summary": "Updates personal information by ID.",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "The ID of the personal information to update.",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "The updated personal information.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonalInfoRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonalInfoRequestDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PersonalInfoRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PersonalInfoResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonalInfoResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonalInfoResponseDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["PersonalInfo"],
        "summary": "Deletes personal information by ID.",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "The ID of the personal information to delete.",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PersonalInfoResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonalInfoResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonalInfoResponseDto"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CitizenshipResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CityResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DisabilityResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "MaritalStatusResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PersonalInfoRequestDto": {
        "type": "object",
        "properties": {
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "patronymic": {
            "type": "string",
            "nullable": true
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date-time"
          },
          "passportSeries": {
            "type": "string",
            "nullable": true
          },
          "passportNumber": {
            "type": "string",
            "nullable": true
          },
          "passportIssuedBy": {
            "type": "string",
            "nullable": true
          },
          "passportIssueDate": {
            "type": "string",
            "format": "date-time"
          },
          "identificationNumber": {
            "type": "string",
            "nullable": true
          },
          "placeOfBirth": {
            "type": "string",
            "nullable": true
          },
          "currentCityId": {
            "type": "integer",
            "format": "int32"
          },
          "currentAddress": {
            "type": "string",
            "nullable": true
          },
          "homePhone": {
            "type": "string",
            "nullable": true
          },
          "mobilePhone": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "workplace": {
            "type": "string",
            "nullable": true
          },
          "position": {
            "type": "string",
            "nullable": true
          },
          "registrationCityId": {
            "type": "integer",
            "format": "int32"
          },
          "maritalStatusId": {
            "type": "integer",
            "format": "int32"
          },
          "citizenshipId": {
            "type": "integer",
            "format": "int32"
          },
          "disabilityId": {
            "type": "integer",
            "format": "int32"
          },
          "isPensioner": {
            "type": "boolean"
          },
          "monthlyIncome": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "PersonalInfoResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "patronymic": {
            "type": "string",
            "nullable": true
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date-time"
          },
          "passportSeries": {
            "type": "string",
            "nullable": true
          },
          "passportNumber": {
            "type": "string",
            "nullable": true
          },
          "passportIssuedBy": {
            "type": "string",
            "nullable": true
          },
          "passportIssueDate": {
            "type": "string",
            "format": "date-time"
          },
          "identificationNumber": {
            "type": "string",
            "nullable": true
          },
          "placeOfBirth": {
            "type": "string",
            "nullable": true
          },
          "currentCityId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "currentCity": {
            "type": "string",
            "nullable": true
          },
          "currentAddress": {
            "type": "string",
            "nullable": true
          },
          "homePhone": {
            "type": "string",
            "nullable": true
          },
          "mobilePhone": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "workplace": {
            "type": "string",
            "nullable": true
          },
          "position": {
            "type": "string",
            "nullable": true
          },
          "registrationCityId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "registrationCity": {
            "type": "string",
            "nullable": true
          },
          "maritalStatusId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "maritalStatus": {
            "type": "string",
            "nullable": true
          },
          "citizenshipId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "citizenship": {
            "type": "string",
            "nullable": true
          },
          "disabilityId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "disability": {
            "type": "string",
            "nullable": true
          },
          "isPensioner": {
            "type": "boolean"
          },
          "monthlyIncome": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      }
    }
  }
}